// inter/byosnap-node-ts/src/snapser-internal/model/authAnonLoginResponse.ts
/**
 * juhi-dev
 * Your custom SDK
 *
 * The version of the OpenAPI document: juhi-dev: v4 SDK
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { AuthUser } from './authUser';

export class AuthAnonLoginResponse {
    'user'?: AuthUser;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "user",
            "baseName": "user",
            "type": "AuthUser"
        }    ];

    static getAttributeTypeMap() {
        return AuthAnonLoginResponse.attributeTypeMap;
    }
}