// inter/byosnap-node-ts/src/snapser-internal/model/authAppleLoginRequest.ts
/**
 * juhi-dev
 * Your custom SDK
 *
 * The version of the OpenAPI document: juhi-dev: v4 SDK
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { AuthLoginMetadata } from './authLoginMetadata';

export class AuthAppleLoginRequest {
    /**
    * Whether to create a user, if it does not exist
    */
    'createUser'?: boolean;
    'loginMetadata'?: AuthLoginMetadata;
    /**
    * Apple token generated on the client
    */
    'token': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "createUser",
            "baseName": "create_user",
            "type": "boolean"
        },
        {
            "name": "loginMetadata",
            "baseName": "login_metadata",
            "type": "AuthLoginMetadata"
        },
        {
            "name": "token",
            "baseName": "token",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AuthAppleLoginRequest.attributeTypeMap;
    }
}