// inter/byosnap-node-ts/src/snapser-internal/model/authAssociateLoginsRequest.ts
/**
 * juhi-dev
 * Your custom SDK
 *
 * The version of the OpenAPI document: juhi-dev: v4 SDK
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class AuthAssociateLoginsRequest {
    /**
    * Token of the user to discard
    */
    'discardUserToken': string;
    /**
    * Token of the user to keep
    */
    'keepUserToken': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "discardUserToken",
            "baseName": "discard_user_token",
            "type": "string"
        },
        {
            "name": "keepUserToken",
            "baseName": "keep_user_token",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AuthAssociateLoginsRequest.attributeTypeMap;
    }
}