// inter/byosnap-node-ts/src/snapser-internal/model/authDiscordLoginRequest.ts
/**
 * juhi-dev
 * Your custom SDK
 *
 * The version of the OpenAPI document: juhi-dev: v4 SDK
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { AuthLoginMetadata } from './authLoginMetadata';

export class AuthDiscordLoginRequest {
    /**
    * Discord access token generated on the client. One of code or access_token is required
    */
    'accessToken'?: string;
    /**
    * Discord code generated on the client. One of code or access_token is required
    */
    'code'?: string;
    /**
    * Whether to create a user, if it does not exist
    */
    'createUser'?: boolean;
    'loginMetadata'?: AuthLoginMetadata;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "accessToken",
            "baseName": "access_token",
            "type": "string"
        },
        {
            "name": "code",
            "baseName": "code",
            "type": "string"
        },
        {
            "name": "createUser",
            "baseName": "create_user",
            "type": "boolean"
        },
        {
            "name": "loginMetadata",
            "baseName": "login_metadata",
            "type": "AuthLoginMetadata"
        }    ];

    static getAttributeTypeMap() {
        return AuthDiscordLoginRequest.attributeTypeMap;
    }
}