// inter/byosnap-node-ts/src/snapser-internal/model/authEmailPasswordLoginRequest.ts
/**
 * my-dev-cluster
 * Your custom SDK
 *
 * The version of the OpenAPI document: my-dev-cluster: v1 SDK
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class AuthEmailPasswordLoginRequest {
    /**
    * Whether to create a user, if it does not exist
    */
    'createUser'?: boolean;
    /**
    * Email address used as login
    */
    'email': string;
    /**
    * Password used for login
    */
    'password': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "createUser",
            "baseName": "create_user",
            "type": "boolean"
        },
        {
            "name": "email",
            "baseName": "email",
            "type": "string"
        },
        {
            "name": "password",
            "baseName": "password",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AuthEmailPasswordLoginRequest.attributeTypeMap;
    }
}