// inter/byosnap-node-ts/src/snapser-internal/model/authEmailPasswordLoginResponse.ts
/**
 * my-dev-cluster
 * Your custom SDK
 *
 * The version of the OpenAPI document: my-dev-cluster: v1 SDK
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { AuthUser } from './authUser';

export class AuthEmailPasswordLoginResponse {
    /**
    * Whether email verification is required or not
    */
    'emailVerificationRequired'?: boolean;
    'user'?: AuthUser;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "emailVerificationRequired",
            "baseName": "email_verification_required",
            "type": "boolean"
        },
        {
            "name": "user",
            "baseName": "user",
            "type": "AuthUser"
        }    ];

    static getAttributeTypeMap() {
        return AuthEmailPasswordLoginResponse.attributeTypeMap;
    }
}